+++ include/abstract_index.h
@@ -66,13 +66,7 @@ class AbstractIndex
 
     // Initialize space for res_vectors before calling.
     template <typename data_type, typename tag_type>
-    size_t search_with_callback(const data_type *query, const uint64_t K, const uint32_t L, tag_type *tags, float *distances, std::vector<data_type *> &res_vectors, const std::function<bool(const uint32_t&, float&)> callback);
-
-
-    //// Initialize space for res_vectors before calling.
-    //template <typename data_type, typename tag_type>
-    //size_t search_with_callback(const data_type *query, const uint64_t K, const uint32_t L, tag_type *tags,
-    //                        float *distances, std::vector<data_type *> &res_vectors, const std::function<bool(const uint32_t &, float &)> &callback);
+    size_t search_with_callback(const data_type *query, const uint64_t K, const uint32_t L, tag_type *tags, float *distances, std::vector<data_type *> &res_vectors, const std::function<bool(const int64_t&, float&)> callback);
 
     // Added search overload that takes L as parameter, so that we
     // can customize L on a per-query basis without tampering with "Parameters"
@@ -132,7 +126,7 @@ class AbstractIndex
     virtual size_t _search_with_tags(const DataType &query, const uint64_t K, const uint32_t L, const TagType &tags,
                                      float *distances, DataVector &res_vectors) = 0;
     virtual size_t _search_with_callback(const DataType &query, const uint64_t K, const uint32_t L, const TagType &tags,
-                                         float *distances, DataVector &res_vectors, const std::function<bool(const uint32_t&, float&)> callback) = 0;
+                                         float *distances, DataVector &res_vectors, const std::function<bool(const int64_t&, float&)> callback) = 0;
 
     //virtual size_t _search_with_callback(const DataType &query, const uint64_t K, const uint32_t L, const TagType &tags,
     //                                 float *distances, DataVector &res_vectors, const std::function<bool(const uint32_t &, float &)> &callback) = 0;
